[Refactor] 급여 도메인 구조 개편 및 관리자 영역 추가
