+++ hero-fe/src/api/payroll/payroll.admin.ts
@@ -1,0 +1,26 @@
+/**
+ * TypeScript Name   : payroll.admin.ts
+ * Description : 급여(Payroll) 관리자용 API 호출 모듈
+ *               - 급여 대시보드
+ *               - 월별 급여 배치
+ *               - 수당 / 공제 항목 관리(CRUD)
+ *               - 급여 조정
+ *               - 사원 급여 조회
+ *               - 급여 명세서 관리 (명세서 리스트)
+ *               - 급여 지급 이력
+ *               - 급여 보고서
+ *               - 급여 설정
+ * History
+ * 2025/12/12 - 동근 최초 작성
+ * </pre>
+ *
+ * @author 동근
+ * @version 1.0
+ */
+import client from '@/api/apiClient';
+
+// import type {
+
+// } from '@/types/payroll/payroll.admin';
+
+/* ------- 관리자용 급여 API ------- */
