+++ hero-fe/src/router/modules/payrollAdmin.ts
@@ -1,0 +1,15 @@
+/**
+ * <pre>
+ * TypeScript Name   : payrollAdmin.ts
+ * Description : 급여(Payroll) 도메인의 라우트 설정 파일
+ * 
+ *
+ * History
+ * 2025/12/09 - 동근 최초 작성 (vue-router 라우트 구성 [내 급여, 내 급여 이력])
+ * </pre>
+ *
+ * @module payrollAdmin-routes
+ * @author 동근
+ * @version 1.0
+ */
+import type { RouteRecordRaw } from 'vue-router';
