+++ hero-fe/src/stores/payroll/payrollAdminStore.ts
@@ -1,0 +1,24 @@
+/**
+ * <pre>
+ * TypeScript Name : payrollAdminStore.ts
+ * Description     : 급여(Payroll) 도메인의 Pinia Store (프론트 상태 + API 연동)
+ *
+ * Responsibility
+ *   - 
+ *   - 
+ *   - 
+ *   - 
+ *   - 
+ *
+ * Views
+ *   - /payroll               내 급여
+ *   - /payroll/history       내 급여 이력
+ *
+ * History
+ *   2025/12/09 - 동근 최초 작성 (급여 관련 API 연동 + Pinia 상태 관리 구성)
+ * </pre>
+ *
+ * @module payroll-store
+ * @author 동근
+ * @version 1.0
+ */
