+++ hero-fe/src/types/payroll/payroll.admin.ts
@@ -1,0 +1,12 @@
+/**
+ * <pre>
+ * TypeScript Name   : payroll.admin.ts
+ * Description : 
+ *
+ * History
+ * 2025/12/12 - 동근 최초 작성
+ * </pre>
+ *
+ * @author 동근
+ * @version 1.0
+ */
