+++ hero-fe/src/views/payroll/admin/Adjust.vue
@@ -1,0 +1,14 @@
+<!-- 
+ * <pre>
+ * Vue Name : Adjust.vue
+ * Description     : 급여 조정(관리자)
+ *
+ *
+ * History
+ *   2025/12/09 - 동근 최초 작성
+ * </pre>
+ *
+ * @module payroll-adjust
+ * @author 동근
+ * @version 1.0
+ -->
