+++ hero-fe/src/views/payroll/admin/Batch.vue
@@ -1,0 +1,14 @@
+<!-- 
+ * <pre>
+ * Vue Name : Batch.vue
+ * Description     : 월별 급여 배치(관리자)
+ *
+ *
+ * History
+ *   2025/12/09 - 동근 최초 작성
+ * </pre>
+ *
+ * @module payroll-batch
+ * @author 동근
+ * @version 1.0
+ -->
