+++ hero-fe/src/views/payroll/admin/Index.vue
@@ -1,0 +1,14 @@
+<!-- 
+ * <pre>
+ * Vue Name : Index.vue
+ * Description     : 급여 대시보드(관리자)
+ *
+ *
+ * History
+ *   2025/12/09 - 동근 최초 작성
+ * </pre>
+ *
+ * @module payroll-dashboard
+ * @author 동근
+ * @version 1.0
+ -->
