+++ hero-fe/src/views/payroll/admin/Items.vue
@@ -1,0 +1,14 @@
+<!-- 
+ * <pre>
+ * Vue Name : Items.vue
+ * Description     : 급여 항목 관리(수당/공제수당)(관리자)
+ *
+ *
+ * History
+ *   2025/12/09 - 동근 최초 작성
+ * </pre>
+ *
+ * @module payroll-items
+ * @author 동근
+ * @version 1.0
+ -->
