+++ hero-fe/src/views/payroll/admin/PaymentHistory.vue
@@ -1,0 +1,14 @@
+<!-- 
+ * <pre>
+ * Vue Name : PaymentHistory.vue
+ * Description     : 급여 지급 이력(관리자)
+ *
+ *
+ * History
+ *   2025/12/09 - 동근 최초 작성
+ * </pre>
+ *
+ * @module payroll-payment-history
+ * @author 동근
+ * @version 1.0
+ -->
