+++ hero-fe/src/views/payroll/admin/Policy.vue
@@ -1,0 +1,14 @@
+<!-- 
+ * <pre>
+ * Vue Name : Policy.vue
+ * Description     : 급여 설정(관리자)
+ *
+ *
+ * History
+ *   2025/12/09 - 동근 최초 작성
+ * </pre>
+ *
+ * @module payroll-policy
+ * @author 동근
+ * @version 1.0
+ -->
