+++ hero-fe/src/views/payroll/admin/Report.vue
@@ -1,0 +1,14 @@
+<!-- 
+ * <pre>
+ * Vue Name : Report.vue
+ * Description     : 급여 보고서(관리자)
+ *
+ *
+ * History
+ *   2025/12/09 - 동근 최초 작성
+ * </pre>
+ *
+ * @module payroll-report
+ * @author 동근
+ * @version 1.0
+ -->
