+++ hero-fe/src/views/payroll/admin/Search.vue
@@ -1,0 +1,14 @@
+<!-- 
+ * <pre>
+ * Vue Name : Search.vue
+ * Description     : 사원 급여 조회(관리자)
+ *
+ *
+ * History
+ *   2025/12/09 - 동근 최초 작성
+ * </pre>
+ *
+ * @module payroll-search
+ * @author 동근
+ * @version 1.0
+ -->
